Store cached values on components (#22)
Currently, all cached values are stored in a global hashmap, which means that these cached values will never get marked for disposal by the garbage collector.

This introduces a CachedValueDataHolder, which allows cached values to be written to it. This allows these cached values to be disposed of when the CacheValueDataHolder itself is disposed.